refactor(home): add cards folder
