+++ FinEdgeFrontend/src/Components/HomePage/Cards/InfoCard.tsx
