+++ FinEdgeFrontend/src/Components/HomePage/Cards/StatCard.tsx
