+++ FinEdgeFrontend/src/Pages/Accounts.tsx
@@ -9,9 +9,9 @@ const Accounts = () => {
       <div className="p-6 bg-gray-50">
         <div className="space-y-6">
           <div className="gap-4 grid lg:grid-cols-3 md:grid-cols-1">
-              <CreateAccounts />
-              <AccountPieChart />
-              <QuickStats />
+            <CreateAccounts />
+            <AccountPieChart />
+            <QuickStats />
           </div>
           <div className="bg-white rounded-lg shadow">
             <div className="p-4">
